pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlet-tc05"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "databricks-connect>=15.4,<16",
    "databricks-sdk>=0.105.0",
    "fastapi>=0.136.1",
    "joblib>=1.5.3",
    "loguru>=0.7.3",
    "matplotlib>=3.10.9",
    "mlflow>=2.0.0",
    "numpy>=1.26,<2",
    "pandas>=2.2,<3",
    "pydantic>=2.13.3",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.8.0",
    "tensorflow>=2.21.0",
    "uvicorn>=0.46.0",
    "yfinance>=1.3.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.0",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=finance --cov-report=term-missing --cov-fail-under=60"

[tool.coverage.run]
omit = [
    "src/finance/models/*",
]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
